Move @devhints/helpers to src/helpers/
