--- packages/helpers/package.json
@@ -1,5 +1,0 @@
-{
-  "name": "@devhints/helpers",
-  "description": "Private helpers and stuff",
-  "version": "0.0.0"
-}
